Bug fixes for race Angelscript, AI checkpoint collisions. Player model read from ".champ" file.
The raceList in races.as is replaced with a dictionary lookup now to fully support ChampionshipManager
generated raceIDs. Logging has been added for the race result logic on the Race / Championship  selectors. This is
buggy and needs overhauling. RaceResults and RaceCompetitors need to be stored as pointers in the vectors instead of
by value as references to them seem to be changing.